% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_package_tmp}
\alias{install_package_tmp}
\title{Download, extract and install a specific version of a package}
\usage{
install_package_tmp(package, version, quiet = TRUE, source = NULL)
}
\arguments{
\item{package}{Package name, character scalar.}

\item{version}{Version number, character scalar. If \code{NULL} then
the latest available version is used.}

\item{quiet}{Whether to show the output of the download/install
process.}
}
\value{
A path to a directory where the package is downloaded,
  extracted and installed. Subdirectory \code{src/<package>}
  contains the extracted source package, \code{src} also contained
  the tarball. Subdirectory \code{lib} contains the library with
  the installed package(s). The dependencies are also installed here.
}
\description{
Download, extract and install a specific version of a package
}
\keyword{internal}

