% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{is_on_github}
\alias{is_on_github}
\title{Try to find out whether the package is on GitHub}
\usage{
is_on_github(package, version = NULL)
}
\arguments{
\item{package}{Package name.}

\item{version}{Package version, ignored currently.}
}
\value{
\code{1} if the package is on GitHub,
  \code{0} if we could not decide.
}
\description{
We try to look for a \code{URL} and \code{BugReports}
field in the \code{DESCRIPTION} file, and check if
they point to GitHub.
}
\keyword{internal}
