% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{list_package_metrics}
\alias{list_package_metrics}
\title{List all implemented metrics}
\usage{
list_package_metrics()
}
\value{
A named character vector of metrics descriptions.
  The named are the three letter metrics ids.
}
\description{
List all implemented metrics
}
