% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{package_metrics_csv}
\alias{package_metrics_csv}
\title{Calculate metrics and write them to a CSV file}
\usage{
package_metrics_csv(package, file, version = NULL,
  metrics = names(list_package_metrics()), source = NULL)
}
\arguments{
\item{package}{Package name, character scalar.}

\item{file}{Output file name.}

\item{version}{Version number, character scalar. If \code{NULL},
then the latest version is used.}

\item{metrics}{The ids of the metrics to calculate. The default
is to calculate all implemented metrics.}

\item{source}{Where to get the package sources. CRAN and BioC
packages are found automatically, others need to be specified.}
}
\description{
Calculate metrics and write them to a CSV file
}
\seealso{
\code{\link{package_metrics}} to calculate the metrics,
without writing them to a file.
}
