% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{packageMetrics2}
\alias{packageMetrics2}
\alias{packageMetrics2-package}
\title{Collect Metrics about R Packages}
\description{
This package is used to collect metrics about CRAN, BioConductor or
other packages, for Valid-R.
}
